Little code cleanup
